# apps/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "description": "Transaction API for client transactions",
    "title": "Transaction API",
    "version": "1.0.0"
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "paths": {
...
    "/api/v1/users/{userId}/transaction-overrides/{transactionId}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "transactionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "errors": {
                      "items": {},
                      "type": "array"
                    },
                    "retry_after": {
                      "type": "number"
                    },
                    "status": {
                      "type": "number"
                    },
                    "title": {
                      "type": "string"
                    },
                    "trace_id": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "errors": {
                      "items": {},
                      "type": "array"
                    },
                    "retry_after": {
                      "type": "number"
                    },
                    "status": {
                      "type": "number"
                    },
                    "title": {
                      "type": "string"
                    },
                    "trace_id": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "transactions"
        ]
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "transactionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "categoryId": {
                    "exclusiveMinimum": true,
                    "maximum": 9007199254740991,
                    "minimum": 0,
                    "type": "integer"
                  }
                },
                "required": [
                  "categoryId"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "categoryId": {
                      "maximum": 9007199254740991,
                      "minimum": -9007199254740991,
                      "type": "integer"
                    },
                    "isOverridden": {
                      "type": "boolean"
                    },
                    "transactionId": {
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                      "type": "string"
                    },
                    "updatedAt": {
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                      "type": "string"
                    }
                  },
                  "required": [
                    "transactionId",
                    "categoryId",
                    "isOverridden",
                    "updatedAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "errors": {
                      "items": {},
                      "type": "array"
                    },
                    "retry_after": {
                      "type": "number"
                    },
                    "status": {
                      "type": "number"
                    },
                    "title": {
                      "type": "string"
                    },
                    "trace_id": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "errors": {
                      "items": {},
                      "type": "array"
                    },
                    "retry_after": {
                      "type": "number"
                    },
                    "status": {
                      "type": "number"
                    },
                    "title": {
                      "type": "string"
                    },
                    "trace_id": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "detail": {
                      "type": "string"
                    },
                    "errors": {
                      "items": {},
                      "type": "array"
                    },
                    "retry_after": {
                      "type": "number"
                    },
                    "status": {
                      "type": "number"
                    },
                    "title": {
                      "type": "string"
                    },
                    "trace_id": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "transactions"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}